requirements.txt
black==23.7.0
redis==5.0.0
Pillow==10.0.0
loguru==0.7.0
requests==2.31.0
python-dotenv~=1.0.0
cuid2~=2.0.0
requests-cache==1.1.0
nsfw_detector==1.1.1
protobuf==3.20.*
blurhash-python==1.2.1
